Build the kernel and ramdisk after packages.
This fixes a bug with LVM builds that I encountered back in 2010. It's been awhile so I don't recall if this breaks lower stage builds or not, but I know that as-is, the LVM utilities are missing from the kernel and it fails to boot. This was the patch I used at ServerBeach to get around that.